Remove unused third-party evidence connectors
The OpenLIT/OTLP, Paperclip, and Entire Git ingestion connectors had no
producer anywhere in the product: onboard, the managed daemon, the hook
capture path, and the receipt pipeline never construct connector evidence,
and the macOS app and TUI never call their routes. They were reachable only
by a user manually invoking `agentacct connector ...` or POSTing to the
localhost import routes.

Removed:
- connectors/{openlit,paperclip,entire,registry}.py and connector_runtime.py
- the `connector` CLI sub-app (list/paperclip/openlit/entire)
- POST /v1/traces and /connectors/{openlit,paperclip,entire}/import routes
  (plus the now-unused _bounded_json_body / _import_connector helpers and the
  EntireImportRequest model)
- the OpenLIT/Paperclip/Entire rows from the usage truth-table
- connectors/UPSTREAM.md and the connector test fixtures

Kept: the advisory control-signal evaluator (connectors/control.py, the
`control evaluate` command, and POST /control/evaluate), which is unrelated to
ingestion. connectors/base.py is reduced to the JSON-digest helpers the
evaluator still uses. The control-signal tests are preserved by moving them
out of the deleted connector test files into test_control_signal.py (unit),
test_control_api.py (HTTP), and test_control_signal_cli.py (CLI).

The inert source-authority rules for the otlp/git_checkpoint source types
(source_policy.py) are intentionally left in place; they have no producer now
but are covered by the evidence-v2 authority tests and are out of scope here.
