+++ src/agentacct/connectors/__init__.py
@@ -1,16 +1,9 @@
-"""Read-only third-party evidence adapters.
+"""Read-only advisory control-signal evaluation.
 
-No connector in this package owns upstream scheduling, tracing storage, Git
-refs, or external control.  They normalize bounded snapshots into
-``ConnectorRecord`` objects for the evidence kernel.
+The control-signal evaluator scores a single advisory signal against locally
+stored evidence and never dispatches an external action.
 """
 
-from .base import (
-    ConnectorError,
-    ConnectorRecord,
-    EvidenceCoreUnavailable,
-    ReadOnlyConnector,
-)
 from .control import (
     ControlDecision,
     ControlSignal,
@@ -21,25 +14,12 @@ from .control import (
     require_hard_enforcement,
     validate_supporting_evidence,
 )
-from .entire import EntireGitConnector
-from .openlit import OpenLITOTLPConnector
-from .paperclip import PaperclipSnapshotConnector
-from .registry import ConnectorRegistry, build_default_registry
 
 __all__ = [
-    "ConnectorError",
-    "ConnectorRecord",
-    "ConnectorRegistry",
     "ControlDecision",
     "ControlSignal",
-    "EntireGitConnector",
-    "EvidenceCoreUnavailable",
     "HardEnforcementRefused",
-    "OpenLITOTLPConnector",
-    "PaperclipSnapshotConnector",
-    "ReadOnlyConnector",
     "SupportingEvidenceValidation",
-    "build_default_registry",
     "evaluate_control_signal",
     "normalize_supporting_evidence_ids",
     "require_hard_enforcement",
