+++ src/agentacct/usage_truth.py
@@ -238,70 +238,6 @@ USAGE_TRUTH_TABLE: tuple[UsageTruthRow, ...] = (
         ],
     ),
     UsageTruthRow(
-        integration="OpenLIT / OTLP trace ingestion",
-        tier="telemetry",
-        evidence_source="OTLP/HTTP JSON spans exported by OpenLIT or another local OpenTelemetry-compatible producer",
-        update_timing="As spans are exported to the local /v1/traces receiver, or when an OTLP JSON file is imported.",
-        observable_fields=[
-            "trace/span/session identifiers",
-            "runtime lifecycle and tool-span metadata",
-            "allowlisted model/provider labels",
-            "telemetry-reported token or cost fields when present",
-        ],
-        usage_confidence="telemetry_reported; corroborating until its measurement provenance is independently verified",
-        cost_confidence="telemetry_reported; never provider_billed merely because it arrived over OTLP",
-        hard_budget_basis="Advisory only unless a separate provider-billed source proves the cost basis.",
-        automatic_scope="Only spans intentionally exported or imported into agentacct's local receiver.",
-        setup_path="agentacct connector openlit /path/to/otlp.json --dry-run --json",
-        limitations=[
-            "OTLP is a transport format, not proof of billing authority.",
-            "agentacct discards prompt, response, thought, tool argument, and tool result attributes.",
-            "Coverage depends on what the instrumented runtime emits.",
-        ],
-    ),
-    UsageTruthRow(
-        integration="Paperclip snapshot connector",
-        tier="orchestrator_claim",
-        evidence_source="User-supplied Paperclip exported JSON snapshot",
-        update_timing="Only when a snapshot is explicitly imported.",
-        observable_fields=[
-            "company/agent/work-item identifiers",
-            "execution assignment and status claims",
-            "work-product claims",
-            "orchestrator-reported cost claim or explicit missing value",
-        ],
-        usage_confidence=USAGE_UNKNOWN,
-        cost_confidence="orchestrator_claim; zero and missing remain distinct",
-        hard_budget_basis="Advisory only; agentacct never dispatches an external Paperclip action.",
-        automatic_scope="Read-only parsing of the supplied export; no Paperclip API mutation or credential is used.",
-        setup_path="agentacct connector paperclip /path/to/snapshot.json --dry-run --json",
-        limitations=[
-            "Orchestrator claims are not provider invoices or objective machine results.",
-            "Connector does not copy Paperclip's scheduler, approvals, RBAC, or control plane.",
-        ],
-    ),
-    UsageTruthRow(
-        integration="Entire Git checkpoint connector",
-        tier="artifact",
-        evidence_source="Known public Entire checkpoint refs, commit trailers, and allowlisted metadata.json scalars",
-        update_timing="Only when a local repository is explicitly scanned.",
-        observable_fields=[
-            "commit/checkpoint/session identifiers",
-            "public ref and parent count",
-            "allowlisted change-count metadata",
-            "checkpoint-reported usage fields when present",
-        ],
-        usage_confidence="checkpoint_reported; corroborating only",
-        cost_confidence=COST_UNKNOWN,
-        hard_budget_basis="No hard budget basis; Git proves artifact existence, not billing or task completion.",
-        automatic_scope="Read-only Git plumbing against known Entire refs; refs, index, and worktree are never changed.",
-        setup_path="agentacct connector entire /path/to/repo --dry-run --json",
-        limitations=[
-            "Does not ingest prompts, transcripts, messages, or private checkpoint bodies.",
-            "Line-to-prompt attribution remains heuristic.",
-        ],
-    ),
-    UsageTruthRow(
         integration="provider/API proxy",
         tier="provider_proxy",
         evidence_source="Traffic explicitly routed through agentacct provider forwarding",
