--- tests/fixtures/connectors/entire_checkpoint_metadata.json
@@ -1,21 +1,0 @@
-{
-  "checkpoint_id": "checkpoint-1",
-  "session_id": "session-entire-1",
-  "strategy": "manual-commit",
-  "agent": "example-agent",
-  "model": "example-model",
-  "created_at": "2026-07-13T01:00:00Z",
-  "files_touched": [],
-  "stats": {
-    "lines_added": 0,
-    "lines_removed": 0
-  },
-  "token_usage": {
-    "input_tokens": 21,
-    "output_tokens": 8,
-    "total_tokens": 29
-  },
-  "no_change": true,
-  "summary": "ENTIRE_METADATA_SUMMARY_SECRET",
-  "prompt": "ENTIRE_METADATA_PROMPT_SECRET"
-}
