--- tests/fixtures/connectors/openlit_otlp.json
@@ -1,92 +1,0 @@
-{
-  "resourceSpans": [
-    {
-      "resource": {
-        "attributes": [
-          {"key": "service.name", "value": {"stringValue": "coding-agent"}},
-          {"key": "service.version", "value": {"stringValue": "1.2.3"}},
-          {"key": "user.email", "value": {"stringValue": "OPENLIT_RESOURCE_SECRET@example.invalid"}}
-        ]
-      },
-      "scopeSpans": [
-        {
-          "scope": {"name": "openlit"},
-          "spans": [
-            {
-              "traceId": "11111111111111111111111111111111",
-              "spanId": "aaaaaaaaaaaaaaaa",
-              "name": "coding_agent.session",
-              "startTimeUnixNano": "1783904400000000000",
-              "endTimeUnixNano": "1783904405000000000",
-              "attributes": [
-                {"key": "coding_agent.session.id", "value": {"stringValue": "session-1"}},
-                {"key": "coding_agent.client", "value": {"stringValue": "codex"}},
-                {"key": "coding_agent.session.outcome", "value": {"stringValue": "success"}},
-                {"key": "gen_ai.provider.name", "value": {"stringValue": "example-provider"}},
-                {"key": "gen_ai.request.model", "value": {"stringValue": "example-model"}},
-                {"key": "gen_ai.usage.input_tokens", "value": {"intValue": "10"}},
-                {"key": "gen_ai.usage.output_tokens", "value": {"intValue": "5"}},
-                {"key": "gen_ai.usage.cost", "value": {"doubleValue": 0.001}},
-                {"key": "gen_ai.input.messages", "value": {"stringValue": "OPENLIT_PROMPT_SECRET"}},
-                {"key": "gen_ai.output.messages", "value": {"stringValue": "OPENLIT_RESPONSE_SECRET"}},
-                {"key": "http.request.body", "value": {"stringValue": "OPENLIT_BODY_SECRET"}}
-              ]
-            },
-            {
-              "traceId": "11111111111111111111111111111111",
-              "spanId": "aaaaaaaaaaaaaaaa",
-              "name": "coding_agent.session",
-              "startTimeUnixNano": "1783904400000000000",
-              "endTimeUnixNano": "1783904405000000000",
-              "attributes": [
-                {"key": "coding_agent.session.id", "value": {"stringValue": "session-1"}},
-                {"key": "coding_agent.client", "value": {"stringValue": "codex"}},
-                {"key": "coding_agent.session.outcome", "value": {"stringValue": "success"}},
-                {"key": "gen_ai.provider.name", "value": {"stringValue": "example-provider"}},
-                {"key": "gen_ai.request.model", "value": {"stringValue": "example-model"}},
-                {"key": "gen_ai.usage.input_tokens", "value": {"intValue": "10"}},
-                {"key": "gen_ai.usage.output_tokens", "value": {"intValue": "5"}},
-                {"key": "gen_ai.usage.cost", "value": {"doubleValue": 0.001}},
-                {"key": "gen_ai.input.messages", "value": {"stringValue": "OPENLIT_PROMPT_SECRET"}}
-              ]
-            },
-            {
-              "traceId": "11111111111111111111111111111111",
-              "spanId": "bbbbbbbbbbbbbbbb",
-              "parentSpanId": "aaaaaaaaaaaaaaaa",
-              "name": "coding_agent.llm.turn",
-              "startTimeUnixNano": "1783904401000000000",
-              "endTimeUnixNano": "1783904402000000000",
-              "attributes": [
-                {"key": "coding_agent.session.id", "value": {"stringValue": "session-1"}},
-                {"key": "coding_agent.turn.id", "value": {"stringValue": "turn-1"}},
-                {"key": "coding_agent.turn.kind", "value": {"stringValue": "assistant"}},
-                {"key": "gen_ai.usage.total_tokens", "value": {"intValue": "15"}},
-                {"key": "coding_agent.thought", "value": {"stringValue": "OPENLIT_THOUGHT_SECRET"}}
-              ]
-            },
-            {
-              "traceId": "11111111111111111111111111111111",
-              "spanId": "cccccccccccccccc",
-              "parentSpanId": "bbbbbbbbbbbbbbbb",
-              "name": "coding_agent.tool_call",
-              "startTimeUnixNano": "1783904402000000000",
-              "endTimeUnixNano": "1783904403000000000",
-              "status": {"code": 1, "message": "OPENLIT_STATUS_MESSAGE_SECRET"},
-              "attributes": [
-                {"key": "coding_agent.session.id", "value": {"stringValue": "session-1"}},
-                {"key": "coding_agent.turn.id", "value": {"stringValue": "turn-1"}},
-                {"key": "gen_ai.tool.name", "value": {"stringValue": "read_file"}},
-                {"key": "gen_ai.tool.call.id", "value": {"stringValue": "tool-1"}},
-                {"key": "coding_agent.tool.sandboxed", "value": {"boolValue": true}},
-                {"key": "gen_ai.tool.call.arguments", "value": {"stringValue": "OPENLIT_TOOL_ARGUMENT_SECRET"}},
-                {"key": "gen_ai.tool.call.result", "value": {"stringValue": "OPENLIT_TOOL_RESULT_SECRET"}},
-                {"key": "process.command_line", "value": {"stringValue": "OPENLIT_COMMAND_SECRET"}}
-              ]
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
