+++ tests/test_control_signal_cli.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import json
-import os
 from pathlib import Path
 
 from typer.testing import CliRunner
@@ -12,7 +11,6 @@ from agentacct.evidence_runtime import EvidenceRuntime
 
 
 runner = CliRunner()
-FIXTURES = Path(__file__).parent / "fixtures" / "connectors"
 
 
 def _append_cli_billed_evidence(store: Path, *, run_id: str) -> EvidenceEnvelope:
@@ -33,49 +31,6 @@ def _append_cli_billed_evidence(store: Path, *, run_id: str) -> EvidenceEnvelope
     )
     EvidenceRuntime(store).append(envelope)
     return envelope
-
-
-def test_paperclip_cli_imports_claims_and_replay_is_duplicate(tmp_path) -> None:
-    store = tmp_path / "state"
-    args = [
-        "connector",
-        "paperclip",
-        str(FIXTURES / "paperclip_snapshot.json"),
-        "--store-dir",
-        str(store),
-        "--json",
-    ]
-
-    first = runner.invoke(app, args)
-    second = runner.invoke(app, args)
-
-    assert first.exit_code == 0, first.output
-    assert json.loads(first.output)["inserted_count"] == 7
-    assert second.exit_code == 0, second.output
-    assert json.loads(second.output)["duplicate_count"] == 7
-
-
-def test_openlit_cli_dry_run_retains_no_prompt_or_evidence(tmp_path) -> None:
-    store = tmp_path / "state"
-    result = runner.invoke(
-        app,
-        [
-            "connector",
-            "openlit",
-            str(FIXTURES / "openlit_otlp.json"),
-            "--store-dir",
-            str(store),
-            "--dry-run",
-            "--json",
-        ],
-    )
-
-    assert result.exit_code == 0, result.output
-    payload = json.loads(result.output)
-    assert payload["record_count"] == 5
-    assert payload["dry_run"] is True
-    assert not (store / "evidence-v2").exists()
-    assert "OPENLIT_PROMPT_SECRET" not in result.output
 
 
 def test_control_cli_downgrades_untrusted_hard_signal_without_dispatch(tmp_path) -> None:
@@ -216,32 +171,3 @@ def test_control_cli_disabled_evidence_store_fails_closed(tmp_path, monkeypatch)
     assert payload["hard_enforcement_allowed"] is False
     assert payload["supporting_evidence_validation"]["state"] == "unavailable"
     assert payload["external_action_dispatched"] is False
-
-
-def test_connector_list_states_read_only_boundary() -> None:
-    result = runner.invoke(app, ["connector", "list", "--json"])
-
-    assert result.exit_code == 0, result.output
-    payload = json.loads(result.output)
-    assert payload["external_writes"] is False
-    assert {item["name"] for item in payload["connectors"]} == {"paperclip", "openlit", "entire"}
-
-
-def test_connector_cli_disabled_returns_nonzero_without_creating_store(tmp_path) -> None:
-    store = tmp_path / "state"
-    result = runner.invoke(
-        app,
-        [
-            "connector",
-            "openlit",
-            str(FIXTURES / "openlit_otlp.json"),
-            "--store-dir",
-            str(store),
-            "--json",
-        ],
-        env={**os.environ, "AGENT_CHRONICLE_EVIDENCE_V2": "0"},
-    )
-
-    assert result.exit_code == 1
-    assert json.loads(result.output)["errors"] == ["evidence_v2_disabled"]
-    assert not (store / "evidence-v2").exists()
